RR-2553 - Refactor `DeletionReason` mapping and enhance delete screener docs
Introduces explicit mapping between the `DeletionReason` resource model and domain entity to improve API-domain separation for the ALN screener deletion process.

Also clarifies the OpenAPI documentation for the `DELETE /profile/{prisonNumber}/aln-screener` endpoint.
